+++ src/components/MovieDialogs/MovieDialogs.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { MovieForm, MovieFormData } from "../MovieForm/MovieForm";
-import { Dialog } from "../utils/Dialog/Dialog.tsx";
+import { Dialog } from "../shared/Dialog/Dialog.tsx";
 
 interface MovieDialogProps {
   isOpen: boolean;
@@ -39,6 +39,32 @@ interface DeleteMovieDialogProps extends MovieDialogProps {
   onSubmit: () => void;
 }
 
+const DeleteMovieDialogContent: React.FC<
+  Omit<DeleteMovieDialogProps, "isOpen">
+> = ({ movieTitle, onClose, onSubmit }) => {
+  return (
+    <div className="space-y-4">
+      <p className="text-gray-300">
+        Are you sure you want to delete &#34;{movieTitle}&#34;?
+      </p>
+      <div className="flex justify-end space-x-4">
+        <button
+          onClick={onClose}
+          className="px-4 py-2 rounded-md bg-gray-600 text-white hover:bg-gray-500"
+        >
+          Cancel
+        </button>
+        <button
+          onClick={onSubmit}
+          className="px-4 py-2 rounded-md bg-red-600 text-white hover:bg-red-500"
+        >
+          Confirm
+        </button>
+      </div>
+    </div>
+  );
+};
+
 export const DeleteMovieDialog: React.FC<DeleteMovieDialogProps> = ({
   isOpen,
   onClose,
@@ -47,25 +73,11 @@ export const DeleteMovieDialog: React.FC<DeleteMovieDialogProps> = ({
 }) => {
   return (
     <Dialog isOpen={isOpen} onClose={onClose} title="Delete Movie">
-      <div className="space-y-4">
-        <p className="text-gray-300">
-          Are you sure you want to delete &#34;{movieTitle}&#34;?
-        </p>
-        <div className="flex justify-end space-x-4">
-          <button
-            onClick={onClose}
-            className="px-4 py-2 rounded-md bg-gray-600 text-white hover:bg-gray-500"
-          >
-            Cancel
-          </button>
-          <button
-            onClick={onSubmit}
-            className="px-4 py-2 rounded-md bg-red-600 text-white hover:bg-red-500"
-          >
-            Confirm
-          </button>
-        </div>
-      </div>
+      <DeleteMovieDialogContent
+        movieTitle={movieTitle}
+        onSubmit={onSubmit}
+        onClose={onClose}
+      />
     </Dialog>
   );
 };
