+++ src/components/MovieTile/MovieTile.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
-import { Image } from "../utils/Image.tsx";
-import { Genres } from "../utils/Genres.tsx";
+import { Image } from "../shared/Image.tsx";
+import { Genres } from "../shared/Genres.tsx";
 import { Movie } from "../../models/movie.interface.ts";
 
 interface MovieTileProps {
